move most of `shared` into `mainnet`
