+++ src/main-oeth.ts
@@ -11,8 +11,8 @@ import * as strategies from './oeth/processors/strategies'
 import * as vault from './oeth/processors/vault'
 import * as validateOeth from './oeth/validators/validate-oeth'
 import { run } from './processor'
+import { processStatus } from './processor-templates/processor-status'
 import * as exchangeRatesPostProcessor from './shared/post-processors/exchange-rates'
-import { processStatus } from './shared/processor-templates/processor-status'
 
 export const processor = {
   stateSchema: 'oeth-processor',
