+++ src/main-ogv.ts
@@ -3,7 +3,7 @@ import * as governance from './ogv/post-processors/governance'
 import * as ogv from './ogv/processors/ogv'
 import * as ogvSupply from './ogv/processors/ogv-supply'
 import { run } from './processor'
-import { processStatus } from './shared/processor-templates/processor-status'
+import { processStatus } from './processor-templates/processor-status'
 
 export const processor = {
   stateSchema: 'ogv-processor',
