+++ src/mainnet/validators/validate-shared.ts
