+++ src/oeth/processors/balancer-meta-pool.ts
@@ -5,8 +5,8 @@ import * as metaStablePool from '../../abi/balancer-meta-stable-pool'
 import * as baseRewardPool4626 from '../../abi/base-reward-pool-4626'
 import { OETHBalancerMetaPoolStrategy } from '../../model'
 import { Context } from '../../processor'
+import { getBalancerStrategyHoldings } from '../../processor-templates/strategy/strategy-balancer'
 import { ensureExchangeRates } from '../../shared/post-processors/exchange-rates'
-import { getBalancerStrategyHoldings } from '../../shared/processor-templates/strategy/strategy-balancer'
 import { RETH_ADDRESS, WETH_ADDRESS } from '../../utils/addresses'
 import { getLatestEntity } from '../../utils/utils'
 import { oethStrategies } from './strategies'
