+++ src/ousd/processors/strategies/convex-meta-strategy.ts
@@ -1,9 +1,7 @@
-import * as erc20 from '../../../abi/erc20'
 import * as abstractStrategyAbi from '../../../abi/initializable-abstract-strategy'
+import { IStrategyData } from '../../../processor-templates/strategy'
 import { currencies } from '../../../shared/post-processors/exchange-rates/currencies'
-import { IStrategyData } from '../../../shared/processor-templates/strategy'
 import { OUSD_ADDRESS } from '../../../utils/addresses'
-import { logFilter } from '../../../utils/logFilter'
 import { traceFilter } from '../../../utils/traceFilter'
 import { DAI, USDC, USDT } from './const'
