+++ src/ousd/processors/strategies/morpho-compound.ts
@@ -1,5 +1,5 @@
+import { IStrategyData } from '../../../processor-templates/strategy'
 import { currencies } from '../../../shared/post-processors/exchange-rates/currencies'
-import { IStrategyData } from '../../../shared/processor-templates/strategy'
 import { OUSD_ADDRESS } from '../../../utils/addresses'
 import { DAI, USDC, USDT } from './const'
