+++ src/processor-templates/curve/curve.ts
@@ -1,17 +1,17 @@
 import { EvmBatchProcessor } from '@subsquid/evm-processor'
 
-import * as curveLpToken from '../../../abi/curve-lp-token'
+import * as curveLpToken from '../../abi/curve-lp-token'
+import { registerLiquiditySource } from '../../mainnet/processors/liquidity-sources'
 import {
   CurvePool,
   CurvePoolBalance,
   CurvePoolRate,
   LiquiditySourceType,
-} from '../../../model'
-import { Context } from '../../../processor'
-import { blockFrequencyUpdater } from '../../../utils/blockFrequencyUpdater'
-import { range } from '../../../utils/range'
-import { updateLiquidityBalances } from '../../post-processors/liquidity'
-import { registerLiquiditySource } from '../../processors/liquidity-sources'
+} from '../../model'
+import { Context } from '../../processor'
+import { updateLiquidityBalances } from '../../shared/post-processors/liquidity'
+import { blockFrequencyUpdater } from '../../utils/blockFrequencyUpdater'
+import { range } from '../../utils/range'
 
 interface ProcessResult {
   curvePoolBalances: CurvePoolBalance[]
