+++ src/processor-templates/maverick/index.ts
