+++ src/processor-templates/otoken/otoken.ts
@@ -1,11 +1,10 @@
 import { EvmBatchProcessor } from '@subsquid/evm-processor'
 import { groupBy } from 'lodash'
 import { GetTransactionReceiptReturnType } from 'viem'
-import { mainnet } from 'viem/chains'
 
-import * as erc20 from '../../../abi/erc20'
-import * as otoken from '../../../abi/otoken'
-import * as otokenVault from '../../../abi/otoken-vault'
+import * as erc20 from '../../abi/erc20'
+import * as otoken from '../../abi/otoken'
+import * as otokenVault from '../../abi/otoken-vault'
 import {
   HistoryType,
   OETH,
@@ -25,19 +24,19 @@ import {
   OUSDRebase,
   OUSDRebaseOption,
   RebasingOption,
-} from '../../../model'
-import { Context } from '../../../processor'
-import { type Transaction, activityFromTx } from '../../../utils/activityFromTx'
-import { ADDRESS_ZERO } from '../../../utils/addresses'
-import { DECIMALS_18 } from '../../../utils/constants'
-import { multicall } from '../../../utils/multicall'
-import { EntityClassT, InstanceTypeOfConstructor } from '../../../utils/type'
-import { getLatestEntity } from '../../../utils/utils'
-import { ensureExchangeRate } from '../../post-processors/exchange-rates'
+} from '../../model'
+import { Context } from '../../processor'
+import { ensureExchangeRate } from '../../shared/post-processors/exchange-rates'
 import {
   CurrencyAddress,
   CurrencySymbol,
-} from '../../post-processors/exchange-rates/currencies'
+} from '../../shared/post-processors/exchange-rates/currencies'
+import { type Transaction, activityFromTx } from '../../utils/activityFromTx'
+import { ADDRESS_ZERO } from '../../utils/addresses'
+import { DECIMALS_18 } from '../../utils/constants'
+import { multicall } from '../../utils/multicall'
+import { EntityClassT, InstanceTypeOfConstructor } from '../../utils/type'
+import { getLatestEntity } from '../../utils/utils'
 import { createAddress, createRebaseAPY } from './utils'
 
 type OToken = EntityClassT<OETH> | EntityClassT<OUSD>
