+++ src/processor-templates/strategy-rewards/index.ts
