+++ src/processor-templates/strategy-rewards/strategy-rewards.ts
@@ -1,9 +1,9 @@
 import { EvmBatchProcessor } from '@subsquid/evm-processor'
 
-import * as iat from '../../../abi/initializable-abstract-strategy'
-import { OETHRewardTokenCollected } from '../../../model'
-import { Context } from '../../../processor'
-import { EntityClassT } from '../../../utils/type'
+import * as iat from '../../abi/initializable-abstract-strategy'
+import { OETHRewardTokenCollected } from '../../model'
+import { Context } from '../../processor'
+import { EntityClassT } from '../../utils/type'
 
 export const createStrategyRewardSetup = ({
   address,
