+++ src/processor-templates/strategy/index.ts
