+++ src/processor-templates/strategy/strategy-curve-amo.ts
@@ -1,12 +1,12 @@
 import { EvmBatchProcessor } from '@subsquid/evm-processor'
 
-import * as curvePool from '../../../abi/curve-lp-token'
-import * as erc20 from '../../../abi/erc20'
-import * as abstractStrategyAbi from '../../../abi/initializable-abstract-strategy'
-import { StrategyBalance } from '../../../model'
-import { Block, Context } from '../../../processor'
-import { ETH_ADDRESS, WETH_ADDRESS } from '../../../utils/addresses'
-import { blockFrequencyUpdater } from '../../../utils/blockFrequencyUpdater'
+import * as curvePool from '../../abi/curve-lp-token'
+import * as erc20 from '../../abi/erc20'
+import * as abstractStrategyAbi from '../../abi/initializable-abstract-strategy'
+import { StrategyBalance } from '../../model'
+import { Block, Context } from '../../processor'
+import { ETH_ADDRESS, WETH_ADDRESS } from '../../utils/addresses'
+import { blockFrequencyUpdater } from '../../utils/blockFrequencyUpdater'
 import { IStrategyData } from './index'
 import {
   processStrategyEarnings,
