+++ src/processor-templates/strategy/strategy-daily-earnings.ts
@@ -2,10 +2,10 @@ import dayjs from 'dayjs'
 import { Between, LessThan, LessThanOrEqual } from 'typeorm'
 import { parseEther } from 'viem'
 
-import { StrategyDailyYield, StrategyYield } from '../../../model'
-import { Block, Context } from '../../../processor'
-import { calculateAPY } from '../../../utils/calculateAPY'
-import { lastExcept, max } from '../../../utils/utils'
+import { StrategyDailyYield, StrategyYield } from '../../model'
+import { Block, Context } from '../../processor'
+import { calculateAPY } from '../../utils/calculateAPY'
+import { lastExcept, max } from '../../utils/utils'
 import { IStrategyData } from './strategy'
 
 const eth1 = 1000000000000000000n
