+++ src/processor-templates/strategy/strategy-generic.ts
@@ -1,11 +1,11 @@
 import { EvmBatchProcessor } from '@subsquid/evm-processor'
 
-import * as abstractStrategyAbi from '../../../abi/initializable-abstract-strategy'
-import { StrategyBalance } from '../../../model'
-import { Block, Context } from '../../../processor'
-import { blockFrequencyUpdater } from '../../../utils/blockFrequencyUpdater'
-import { ensureExchangeRates } from '../../post-processors/exchange-rates'
-import { CurrencyAddress } from '../../post-processors/exchange-rates/currencies'
+import * as abstractStrategyAbi from '../../abi/initializable-abstract-strategy'
+import { StrategyBalance } from '../../model'
+import { Block, Context } from '../../processor'
+import { ensureExchangeRates } from '../../shared/post-processors/exchange-rates'
+import { CurrencyAddress } from '../../shared/post-processors/exchange-rates/currencies'
+import { blockFrequencyUpdater } from '../../utils/blockFrequencyUpdater'
 import { IStrategyData } from './index'
 import {
   processStrategyEarnings,
