+++ src/processor-templates/strategy/strategy.ts
@@ -1,13 +1,12 @@
 import { EvmBatchProcessor } from '@subsquid/evm-processor'
-import durationPlugin from 'dayjs/plugin/duration'
 
-import { Context } from '../../../processor'
-import { OETH_ADDRESS, OUSD_ADDRESS } from '../../../utils/addresses'
-import { blockFrequencyUpdater } from '../../../utils/blockFrequencyUpdater'
-import { LogFilter } from '../../../utils/logFilter'
-import { TraceFilter } from '../../../utils/traceFilter'
-import { ensureExchangeRates } from '../../post-processors/exchange-rates'
-import { CurrencyAddress } from '../../post-processors/exchange-rates/currencies'
+import { Context } from '../../processor'
+import { ensureExchangeRates } from '../../shared/post-processors/exchange-rates'
+import { CurrencyAddress } from '../../shared/post-processors/exchange-rates/currencies'
+import { OETH_ADDRESS, OUSD_ADDRESS } from '../../utils/addresses'
+import { blockFrequencyUpdater } from '../../utils/blockFrequencyUpdater'
+import { LogFilter } from '../../utils/logFilter'
+import { TraceFilter } from '../../utils/traceFilter'
 import * as strategyBalancer from './strategy-balancer'
 import * as strategyCurveAMO from './strategy-curve-amo'
 import * as strategyGeneric from './strategy-generic'
